Move individual nodes into their own files
